info : bugfix

+++ scripts/e8257d_signal_generator.py
@@ -118,7 +118,7 @@ class e8257d_controller(object):
             )
         self.sub_onoff = rospy.Subscriber(
             name = '{}_onoff_cmd'.format(node_name),
-            data_class = Float64,
+            data_class = Int32,
             callback = self.callback_onoff,
             queue_size = 1
             )
